packages/dot-agent-kit/pyproject.toml
[project]
name = "dot-agent-kit"
version = "0.1.28"
description = "Kit management for Claude Code"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.1.0",
    "erk-shared",
    "pathspec>=0.12.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.0.0",
    "rich>=13.0",
    "tomli>=2.0",
    "tomli-w>=1.0",
]

[project.optional-dependencies]
nomcp = [
    "fastmcp>=2.0.0",
]

[project.scripts]
dot-agent = "dot_agent_kit.cli:main"

[tool.uv.sources]
erk-shared = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dot_agent_kit"]

[tool.pyright]
include = ["src", "tests"]
pythonVersion = "3.13"
typeCheckingMode = "strict"
reportPrivateUsage = false
